+++ tests/test_info.py
@@ -16,7 +16,12 @@ class TestInfo(unittest.TestCase):
 
     def test_empty(self):
         """Test missing info."""
-        info = Info({})
+        info = Info({
+            'basic': {},
+            'details': {},
+            'access': {},
+            'link': {}
+        })
 
         self.assertIsNone(info.name)
         self.assertIsNone(info.is_dir)
@@ -29,6 +34,8 @@ class TestInfo(unittest.TestCase):
         self.assertIsNone(info.permissions)
         self.assertIsNone(info.user)
         self.assertIsNone(info.group)
+        self.assertIsNone(info.target)
+        self.assertFalse(info.is_link)
 
     def test_access(self):
         info = Info({
@@ -46,6 +53,15 @@ class TestInfo(unittest.TestCase):
         self.assertEqual(info.group, 'devs')
         self.assertEqual(info.uid, 10)
         self.assertEqual(info.gid, 12)
+
+    def test_link(self):
+        info = Info({
+            'link': {
+                'target': 'foo'
+            }
+        })
+        self.assertTrue(info.is_link)
+        self.assertEqual(info.target, 'foo')
 
     def test_basic(self):
         # Check simple file
@@ -115,3 +131,7 @@ class TestInfo(unittest.TestCase):
         info_copy = info.copy()
         self.assertEqual(info.raw, info_copy.raw)
 
+    def test_get(self):
+        info = Info({'baz':{}})
+        self.assertIsNone(info.get('foo', 'bar'))
+        self.assertIsNone(info.get('baz', 'bar'))