+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Lstat info namespace
+- Link info namespace
+- FS.islink method
+- Info.is_link method
+
+
 ## [2.0.7] - 2017-08-06
 
 ### Fixes